Experimental support of HPE LTO5 drive on iokit backend
