+++ src/tape_drivers/freebsd/cam/cam_cmn.c
@@ -60,7 +60,7 @@
 #include <time.h>
 #include <cam/scsi/scsi_message.h>
 #include "ltfs_copyright.h"
-#include "ibm_tape.h"
+#include "vendor_compat.h"
 #include "libltfs/ltfslogging.h"
 #include "cam_cmn.h"
 
@@ -154,7 +154,7 @@ int camtape_ccb2rc(struct camtape_data *softc, union ccb *ccb)
 	return rc;
 }
 
-int camtape_ioctlrc2err(void *device, int fd, struct scsi_sense_data *sense_data, 
+int camtape_ioctlrc2err(void *device, int fd, struct scsi_sense_data *sense_data,
 	int control_cmd, char **msg)
 {
 	union mterrstat errstat;
@@ -841,7 +841,7 @@ int camtape_get_info(void *device, struct tc_drive_info *info)
 	info->channel = 0;
 	info->target  = 0;
 	info->lun     = -1;
-	
+
 	return 0;
 }
