+++ src/tape_drivers/generic/file/filedebug_conf_tc.c
@@ -52,7 +52,7 @@
 #include "libltfs/ltfslogging.h"
 
 #include "tape_drivers/tape_drivers.h"
-#include "tape_drivers/ibm_tape.h"
+#include "tape_drivers/vendor_compat.h"
 
 #include "filedebug_conf_tc.h"
