+++ src/tape_drivers/hp_tape.h
@@ -1,0 +1,80 @@
+/*
+**
+**  OO_Copyright_BEGIN
+**
+**
+**  Copyright 2010, 2020 IBM Corp. All rights reserved.
+**
+**  Redistribution and use in source and binary forms, with or without
+**   modification, are permitted provided that the following conditions
+**  are met:
+**  1. Redistributions of source code must retain the above copyright
+**     notice, this list of conditions and the following disclaimer.
+**  2. Redistributions in binary form must reproduce the above copyright
+**     notice, this list of conditions and the following disclaimer in the
+**  documentation and/or other materials provided with the distribution.
+**  3. Neither the name of the copyright holder nor the names of its
+**     contributors may be used to endorse or promote products derived from
+**     this software without specific prior written permission.
+**
+**  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS ``AS IS''
+**  AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+**  IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+**  ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
+**  LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+**  CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+**  SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+**  INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+**  CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+**  ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+**  POSSIBILITY OF SUCH DAMAGE.
+**
+**
+**  OO_Copyright_END
+**
+*************************************************************************************
+**
+** COMPONENT NAME:  IBM Linear Tape File System
+**
+** FILE NAME:       tape_drivers/hp_tape.h
+**
+** DESCRIPTION:     Definitions of handling HPE tape devices
+**
+** AUTHOR:          Atsushi Abe
+**                  IBM Tokyo Lab., Japan
+**                  piste@jp.ibm.com
+**
+*************************************************************************************
+*/
+
+#include <stdlib.h>
+#include <errno.h>
+
+#include "tape_drivers/spc_op_codes.h"
+#include "tape_drivers/ssc_op_codes.h"
+#include "tape_drivers/tape_drivers.h"
+
+#include "libltfs/ltfslogging.h"
+#include "libltfs/ltfs_error.h"
+
+#ifndef __hp_tape_h
+
+#define __hp_tape_h
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+#define HP_VENDOR_ID "HP"
+
+extern struct error_table hp_tape_errors[];
+
+int hp_tape_init_timeout(struct timeout_tape** table, int type);
+
+extern struct supported_device *hp_supported_drives[];
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif // __hp_tape_h
