+++ src/tape_drivers/netbsd/scsipi-ibmtape/scsipi_ibmtape.c
@@ -64,7 +64,7 @@
 /* Common header of backend */
 #include "reed_solomon_crc.h"
 #include "crc32c_crc.h"
-#include "ibm_tape.h"
+#include "vendor_compat.h"
 
 /* SCSI command handling functions */
 #include "scsipi_scsi_tape.h"
