+++ src/tape_drivers/osx/iokit/iokit.c
@@ -66,7 +66,7 @@
 /* Common header of backend */
 #include "reed_solomon_crc.h"
 #include "crc32c_crc.h"
-#include "ibm_tape.h"
+#include "vendor_compat.h"
 
 /* iokit functions */
 #ifdef VERSION
@@ -387,6 +387,9 @@ static int _take_dump(struct iokit_data *priv, bool capture_unforced)
 	time_t    now;
 	struct tm *tm_now;
 
+	if (priv->vendor == VENDOR_HP)
+		return 0;
+
 	ltfs_profiler_add_entry(priv->profiler, NULL, TAPEBEND_REQ_ENTER(REQ_TC_TAKEDUMPDRV));
 
 	/* Make base filename */
@@ -465,7 +468,7 @@ static int _cdb_read_buffer(void *device, int id, unsigned char *buf, size_t off
 	cdb[7] = (unsigned char)(len >> 8)     & 0xFF;
 	cdb[8] = (unsigned char) len           & 0xFF;
 
-	timeout = ibm_tape_get_timeout(priv->timeouts, cdb[0]);
+	timeout = get_timeout(priv->timeouts, cdb[0]);
 	if (timeout < 0)
 		return -EDEV_UNSUPPORETD_COMMAND;
 
@@ -519,7 +522,7 @@ static int _cdb_force_dump(struct iokit_data *priv)
 	buf[4] = 0x01;
 	buf[5] = 0x60; /* Diag ID */
 
-	timeout = ibm_tape_get_timeout(priv->timeouts, cdb[0]);
+	timeout = get_timeout(priv->timeouts, cdb[0]);
 	if (timeout < 0)
 		return -EDEV_UNSUPPORETD_COMMAND;
 
@@ -563,7 +566,7 @@ static int _cdb_pri(void *device, unsigned char *buf, int size)
 	cdb[7] = (unsigned char)(size >> 8)  & 0xFF;
 	cdb[8] = (unsigned char) size        & 0xFF;
 
-	timeout = ibm_tape_get_timeout(priv->timeouts, cdb[0]);
+	timeout = get_timeout(priv->timeouts, cdb[0]);
 	if (timeout < 0)
 		return -EDEV_UNSUPPORETD_COMMAND;
 
@@ -672,7 +675,7 @@ static int _cdb_pro(void *device,
 	if (sakey)
 		memcpy(buf + 8, sakey, KEYLEN);
 
-	timeout = ibm_tape_get_timeout(priv->timeouts, cdb[0]);
+	timeout = get_timeout(priv->timeouts, cdb[0]);
 	if (timeout < 0)
 		return -EDEV_UNSUPPORETD_COMMAND;
 
@@ -800,9 +803,12 @@ int iokit_open(const char *devname, void **handle)
 	}
 	strncpy(priv->drive_serial, id_data.unit_serial, UNIT_SERIAL_LENGTH - 1);
 
+	/* Convert vendor_id in inq buffer to integer based id */
+	priv->vendor = get_vendor_id(id_data.vendor_id);
+
 	/* Check the drive is supportable */
-	struct supported_device **cur = ibm_supported_drives;
-	while(*cur) {
+	struct supported_device **cur = get_supported_devs(priv->vendor);
+	while(cur && *cur) {
 		if((! strncmp(id_data.vendor_id, (*cur)->vendor_id, strlen((*cur)->vendor_id)) ) &&
 		   (! strncmp(id_data.product_id, (*cur)->product_id, strlen((*cur)->product_id)) ) ) {
 			drive_type = (*cur)->drive_type;
@@ -812,14 +818,15 @@ int iokit_open(const char *devname, void **handle)
 	}
 
 	if(drive_type > 0) {
-		if (!ibm_tape_is_supported_firmware(drive_type, (unsigned char*)id_data.product_rev)) {
+		if (!drive_has_supported_fw(priv->vendor, drive_type, (unsigned char*)id_data.product_rev)) {
 			iokit_release_exclusive_access(&priv->dev);
 			ret = -EDEV_UNSUPPORTED_FIRMWARE;
 			goto free;
-		} else
+		} else {
 			priv->drive_type = drive_type;
+		}
 	} else {
-		ltfsmsg(LTFS_INFO, 30813I, id_data.product_id);
+		ltfsmsg(LTFS_INFO, 30813I, id_data.vendor_id, id_data.product_id);
 		iokit_release_exclusive_access(&priv->dev);
 		ret = -EDEV_DEVICE_UNSUPPORTABLE; /* Unsupported device */
 		goto free;
@@ -841,10 +848,9 @@ int iokit_open(const char *devname, void **handle)
 	priv->info.target  = 0;
 	priv->info.lun     = -1;
 
-	/* Setup IBM tape specific parameters */
-	standard_table = standard_tape_errors;
-	vendor_table   = ibm_tape_errors;
-	ibm_tape_init_timeout(&priv->timeouts, priv->drive_type);
+	/* Setup vendor specific parameters */
+	init_error_table(priv->vendor, &standard_table, &vendor_table);
+	init_timeout(priv->vendor, &priv->timeouts, priv->drive_type);
 
 	/* Register reservation key */
 	ibm_tape_genkey(priv->key);
@@ -1014,7 +1020,7 @@ int iokit_inquiry_page(void *device, unsigned char page, struct tc_inq_page *inq
 	cdb[2] = page;
 	ltfs_u16tobe(cdb + 3, sizeof(inq->data));
 
-	timeout = ibm_tape_get_timeout(priv->timeouts, cdb[0]);
+	timeout = get_timeout(priv->timeouts, cdb[0]);
 	if (timeout < 0)
 		return -EDEV_UNSUPPORETD_COMMAND;
 
@@ -1089,7 +1095,7 @@ int iokit_test_unit_ready(void *device)
 	/* Build CDB */
 	cdb[0] = TEST_UNIT_READY;
 
-	timeout = ibm_tape_get_timeout(priv->timeouts, cdb[0]);
+	timeout = get_timeout(priv->timeouts, cdb[0]);
 	if (timeout < 0)
 		return -EDEV_UNSUPPORETD_COMMAND;
 
@@ -1157,7 +1163,7 @@ static int _cdb_read(void *device, char *buf, size_t size, boolean_t sili)
 	cdb[3] = (size >> 8)  & 0xFF;
 	cdb[4] =  size        & 0xFF;
 
-	timeout = ibm_tape_get_timeout(priv->timeouts, cdb[0]);
+	timeout = get_timeout(priv->timeouts, cdb[0]);
 	if (timeout < 0)
 		return -EDEV_UNSUPPORETD_COMMAND;
 
@@ -1356,7 +1362,7 @@ static int _cdb_write(void *device, uint8_t *buf, size_t size, bool *ew, bool *p
 	cdb[3] = (size >> 8)  & 0xFF;
 	cdb[4] =  size        & 0xFF;
 
-	timeout = ibm_tape_get_timeout(priv->timeouts, cdb[0]);
+	timeout = get_timeout(priv->timeouts, cdb[0]);
 	if (timeout < 0)
 		return -EDEV_UNSUPPORETD_COMMAND;
 
@@ -1477,7 +1483,7 @@ int iokit_writefm(void *device, size_t count, struct tc_position *pos, bool imme
 	cdb[3] = (count >> 8)  & 0xFF;
 	cdb[4] =  count        & 0xFF;
 
-	timeout = ibm_tape_get_timeout(priv->timeouts, cdb[0]);
+	timeout = get_timeout(priv->timeouts, cdb[0]);
 	if (timeout < 0)
 		return -EDEV_UNSUPPORETD_COMMAND;
 
@@ -1552,7 +1558,7 @@ int iokit_rewind(void *device, struct tc_position *pos)
 
 	cdb[0] = REWIND;
 
-	timeout = ibm_tape_get_timeout(priv->timeouts, cdb[0]);
+	timeout = get_timeout(priv->timeouts, cdb[0]);
 	if (timeout < 0)
 		return -EDEV_UNSUPPORETD_COMMAND;
 
@@ -1631,7 +1637,7 @@ int iokit_locate(void *device, struct tc_position dest, struct tc_position *pos)
 	cdb[3]  = (unsigned char)(dest.partition & 0xff);
 	ltfs_u64tobe(cdb + 4, dest.block);
 
-	timeout = ibm_tape_get_timeout(priv->timeouts, cdb[0]);
+	timeout = get_timeout(priv->timeouts, cdb[0]);
 	if (timeout < 0)
 		return -EDEV_UNSUPPORETD_COMMAND;
 
@@ -1723,7 +1729,7 @@ int iokit_space(void *device, size_t count, TC_SPACE_TYPE type, struct tc_positi
 			break;
 	}
 
-	timeout = ibm_tape_get_timeout(priv->timeouts, cdb[0]);
+	timeout = get_timeout(priv->timeouts, cdb[0]);
 	if (timeout < 0)
 		return -EDEV_UNSUPPORETD_COMMAND;
 
@@ -1774,7 +1780,7 @@ static int _cdb_request_sense(void *device, unsigned char *buf, unsigned char si
 	cdb[0] = REQUEST_SENSE;
 	cdb[4] = (unsigned char)size;
 
-	timeout = ibm_tape_get_timeout(priv->timeouts, cdb[0]);
+	timeout = get_timeout(priv->timeouts, cdb[0]);
 	if (timeout < 0)
 		return -EDEV_UNSUPPORETD_COMMAND;
 
@@ -1825,7 +1831,7 @@ int iokit_erase(void *device, struct tc_position *pos, bool long_erase)
 	if (long_erase)
 		cdb[1] = 0x03;
 
-	timeout = ibm_tape_get_timeout(priv->timeouts, cdb[0]);
+	timeout = get_timeout(priv->timeouts, cdb[0]);
 	if (timeout < 0)
 		return -EDEV_UNSUPPORETD_COMMAND;
 
@@ -1899,7 +1905,7 @@ static int _cdb_load_unload(void *device, bool load)
 	if (load)
 		cdb[4] = 0x01;
 
-	timeout = ibm_tape_get_timeout(priv->timeouts, cdb[0]);
+	timeout = get_timeout(priv->timeouts, cdb[0]);
 	if (timeout < 0)
 		return -EDEV_UNSUPPORETD_COMMAND;
 
@@ -1970,10 +1976,17 @@ int iokit_load(void *device, struct tc_position *pos)
 		return ret;
 	}
 
-	priv->cart_type = buf[2];
 	priv->density_code = buf[8];
 
-	ret = ibm_tape_is_supported_tape(priv->cart_type, priv->density_code, &(priv->is_worm));
+	if (priv->vendor == VENDOR_HP) {
+		priv->cart_type = assume_cart_type(priv->density_code);
+		if (buf[2] == 0x01)
+			priv->is_worm = true;
+	} else {
+		priv->cart_type = buf[2];
+	}
+
+	ret = is_supported_tape(priv->cart_type, priv->density_code, &(priv->is_worm));
 	if(ret == -LTFS_UNSUPPORTED_MEDIUM)
 		ltfsmsg(LTFS_INFO, 30831I, priv->cart_type, priv->density_code);
 
@@ -2030,7 +2043,7 @@ int iokit_readpos(void *device, struct tc_position *pos)
 	cdb[0] = READ_POSITION;
 	cdb[1] = 0x06; /* Long Format */
 
-	timeout = ibm_tape_get_timeout(priv->timeouts, cdb[0]);
+	timeout = get_timeout(priv->timeouts, cdb[0]);
 	if (timeout < 0)
 		return -EDEV_UNSUPPORETD_COMMAND;
 
@@ -2109,7 +2122,7 @@ int iokit_setcap(void *device, uint16_t proportion)
 		cdb[0] = SET_CAPACITY;
 		ltfs_u16tobe(cdb + 3, proportion);
 
-		timeout = ibm_tape_get_timeout(priv->timeouts, cdb[0]);
+		timeout = get_timeout(priv->timeouts, cdb[0]);
 		if (timeout < 0)
 			return -EDEV_UNSUPPORETD_COMMAND;
 
@@ -2155,7 +2168,7 @@ int iokit_format(void *device, TC_FORMAT_TYPE format, const char *vol_name, cons
 	cdb[0] = FORMAT_MEDIUM;
 	cdb[2] = format;
 
-	timeout = ibm_tape_get_timeout(priv->timeouts, cdb[0]);
+	timeout = get_timeout(priv->timeouts, cdb[0]);
 	if (timeout < 0)
 		return -EDEV_UNSUPPORETD_COMMAND;
 
@@ -2328,7 +2341,7 @@ static int _cdb_logsense(void *device, const unsigned char page, const unsigned 
 	cdb[3] = subpage;
 	ltfs_u16tobe(cdb + 7, size);
 
-	timeout = ibm_tape_get_timeout(priv->timeouts, cdb[0]);
+	timeout = get_timeout(priv->timeouts, cdb[0]);
 	if (timeout < 0)
 		return -EDEV_UNSUPPORETD_COMMAND;
 
@@ -2388,7 +2401,7 @@ int iokit_modesense(void *device, const unsigned char page, const TC_MP_PC_TYPE 
 	cdb[3] = subpage;
 	ltfs_u16tobe(cdb + 7, size);
 
-	timeout = ibm_tape_get_timeout(priv->timeouts, cdb[0]);
+	timeout = get_timeout(priv->timeouts, cdb[0]);
 	if (timeout < 0)
 		return -EDEV_UNSUPPORETD_COMMAND;
 
@@ -2438,7 +2451,7 @@ int iokit_modeselect(void *device, unsigned char *buf, const size_t size)
 	cdb[1] = 0x10; /* Set PF bit */
 	ltfs_u16tobe(cdb + 7, size);
 
-	timeout = ibm_tape_get_timeout(priv->timeouts, cdb[0]);
+	timeout = get_timeout(priv->timeouts, cdb[0]);
 	if (timeout < 0)
 		return -EDEV_UNSUPPORETD_COMMAND;
 
@@ -2485,7 +2498,7 @@ int iokit_reserve(void *device)
 	/* TODO: Need to use Persistent Reserve */
 	cdb[0] = RESERVE_UNIT6;
 
-	timeout = ibm_tape_get_timeout(priv->timeouts, cdb[0]);
+	timeout = get_timeout(priv->timeouts, cdb[0]);
 	if (timeout < 0)
 		return -EDEV_UNSUPPORETD_COMMAND;
 
@@ -2557,7 +2570,7 @@ int iokit_release(void *device)
 	/* TODO: Need to use Persistent Reserve */
 	cdb[0] = RELEASE_UNIT6;
 
-	timeout = ibm_tape_get_timeout(priv->timeouts, cdb[0]);
+	timeout = get_timeout(priv->timeouts, cdb[0]);
 	if (timeout < 0)
 		return -EDEV_UNSUPPORETD_COMMAND;
 
@@ -2611,7 +2624,7 @@ static int _cdb_prevent_allow_medium_removal(void *device, bool prevent)
 	if (prevent)
 		cdb[4] = 0x01;
 
-	timeout = ibm_tape_get_timeout(priv->timeouts, cdb[0]);
+	timeout = get_timeout(priv->timeouts, cdb[0]);
 	if (timeout < 0)
 		return -EDEV_UNSUPPORETD_COMMAND;
 
@@ -2695,7 +2708,7 @@ int iokit_write_attribute(void *device, const tape_partition_t part,
 	cdb[7] = part;
 	ltfs_u32tobe(cdb + 10, len);
 
-	timeout = ibm_tape_get_timeout(priv->timeouts, cdb[0]);
+	timeout = get_timeout(priv->timeouts, cdb[0]);
 	if (timeout < 0) {
 		free(buffer);
 		return -EDEV_UNSUPPORETD_COMMAND;
@@ -2758,7 +2771,7 @@ int iokit_read_attribute(void *device, const tape_partition_t part,
 	ltfs_u16tobe(cdb + 8, id);
 	ltfs_u32tobe(cdb + 10, len);
 
-	timeout = ibm_tape_get_timeout(priv->timeouts, cdb[0]);
+	timeout = get_timeout(priv->timeouts, cdb[0]);
 	if (timeout < 0)
 		return -EDEV_UNSUPPORETD_COMMAND;
 
@@ -2824,7 +2837,7 @@ int iokit_allow_overwrite(void *device, const struct tc_position pos)
 	cdb[3] = (unsigned char)(pos.partition & 0xff);
 	ltfs_u64tobe(cdb + 4, pos.block);
 
-	timeout = ibm_tape_get_timeout(priv->timeouts, cdb[0]);
+	timeout = get_timeout(priv->timeouts, cdb[0]);
 	if (timeout < 0)
 		return -EDEV_UNSUPPORETD_COMMAND;
 
@@ -2912,12 +2925,16 @@ int iokit_set_default(void *device)
 	}
 
 	/* set logical block protection */
-	if (global_data.crc_checking) {
-		ltfsmsg(LTFS_DEBUG, 30992D, __FUNCTION__, "Setting LBP");
-		ret = _set_lbp(device, true);
+	if (priv->vendor == VENDOR_IBM) {
+		if (global_data.crc_checking) {
+			ltfsmsg(LTFS_DEBUG, 30992D, __FUNCTION__, "Setting LBP");
+			ret = _set_lbp(device, true);
+		} else {
+			ltfsmsg(LTFS_DEBUG, 30992D, __FUNCTION__, "Resetting LBP");
+			ret = _set_lbp(device, false);
+		}
 	} else {
-		ltfsmsg(LTFS_DEBUG, 30992D, __FUNCTION__, "Resetting LBP");
-		ret = _set_lbp(device, false);
+		ret = DEVICE_GOOD;
 	}
 
 	ltfs_profiler_add_entry(priv->profiler, NULL, TAPEBEND_REQ_EXIT(REQ_TC_SETDEFAULT));
@@ -2926,7 +2943,7 @@ int iokit_set_default(void *device)
 
 /**
  * Get cartridge health information
- * @param device a pointer to the ibmtape backend
+ * @param device a pointer to the iokit backend
  * @return 0 on success or a negative value on error
  */
 
@@ -3294,7 +3311,7 @@ static int _cdb_read_block_limits(void *device) {
 	/* Build CDB */
 	cdb[0] = READ_BLOCK_LIMITS;
 
-	timeout = ibm_tape_get_timeout(priv->timeouts, cdb[0]);
+	timeout = get_timeout(priv->timeouts, cdb[0]);
 	if (timeout < 0)
 		return -EDEV_UNSUPPORETD_COMMAND;
 
@@ -3472,6 +3489,14 @@ static const char *_generate_product_name(const char *product_id)
 		}
 	}
 
+	for (i = 0; hp_supported_drives[i]; ++i) {
+		if (strncmp(hp_supported_drives[i]->product_id, product_id,
+					strlen(hp_supported_drives[i]->product_id)) == 0) {
+			product_name = hp_supported_drives[i]->product_name;
+			break;
+		}
+	}
+
 	return product_name;
 }
 
@@ -3588,7 +3613,7 @@ static int _cdb_spin(void *device, const uint16_t sps, unsigned char **buffer, s
 	ltfs_u16tobe(cdb + 2, sps); /* SECURITY PROTOCOL SPECIFIC */
 	ltfs_u32tobe(cdb + 6, len);
 
-	timeout = ibm_tape_get_timeout(priv->timeouts, cdb[0]);
+	timeout = get_timeout(priv->timeouts, cdb[0]);
 	if (timeout < 0)
 		return -EDEV_UNSUPPORETD_COMMAND;
 
@@ -3635,7 +3660,7 @@ int _cdb_spout(void *device, const uint16_t sps,
 	ltfs_u16tobe(cdb + 2, sps); /* SECURITY PROTOCOL SPECIFIC */
 	ltfs_u32tobe(cdb + 6, size);
 
-	timeout = ibm_tape_get_timeout(priv->timeouts, cdb[0]);
+	timeout = get_timeout(priv->timeouts, cdb[0]);
 	if (timeout < 0)
 		return -EDEV_UNSUPPORETD_COMMAND;
 
@@ -4095,7 +4120,7 @@ int iokit_get_block_in_buffer(void *device, uint32_t *block)
 	cdb[0] = READ_POSITION;
 	cdb[1] = 0x08; /* Extended Format */
 
-	timeout = ibm_tape_get_timeout(priv->timeouts, cdb[0]);
+	timeout = get_timeout(priv->timeouts, cdb[0]);
 	if (timeout < 0)
 		return -EDEV_UNSUPPORETD_COMMAND;
 
@@ -4185,8 +4210,11 @@ struct tape_ops iokit_handler = {
 
 struct tape_ops *tape_dev_get_ops(void)
 {
-	standard_table = standard_tape_errors;
-	vendor_table = ibm_tape_errors;
+	if (!standard_table)
+		standard_table = standard_tape_errors;
+	if (!vendor_table)
+		vendor_table = ibm_tape_errors;
+
 	return &iokit_handler;
 }
