+++ src/tape_drivers/tape_drivers.h
@@ -56,6 +56,9 @@
 #include "libltfs/uthash.h"
 #include "libltfs/tape_ops.h"
 
+#include "libltfs/ltfslogging.h"
+#include "libltfs/ltfs_error.h"
+
 #ifndef __tape_drivers_h
 #define __tape_drivers_h
 
@@ -95,6 +98,46 @@ struct error_table {
 	char     *msg;      /**< Description of error */
 };
 
+static inline int _sense2errorcode(uint32_t sense, struct error_table *table, char **msg, uint32_t mask)
+{
+	int rc = -EDEV_UNKNOWN;
+	int i;
+
+	if (msg)
+		*msg = NULL;
+
+	if (!table)
+		return rc;
+
+	if ( (sense & 0xFFFF00) == 0x044000 )
+		sense = 0x044000;
+	else if ( (sense & 0xFFF000) == 0x048000 ) /* 04/8xxx in TS3100/TS3200 */
+		sense = 0x048000;
+	else if ( (sense & 0xFFF000) == 0x0B4100 ) /* 0B/41xx in TS2900 */
+		sense = 0x0B4100;
+
+	if ( (sense & 0x00FF00) >= 0x008000 || (sense & 0x0000FF) >= 0x000080)
+		rc = -EDEV_VENDOR_UNIQUE;
+
+	i = 0;
+	while (table[i].sense != 0xFFFFFF) {
+		if((table[i].sense & mask) == (sense & mask)) {
+			rc  = table[i].err_code;
+			if (msg)
+				*msg = (char*)(table[i].msg);
+			break;
+		}
+		i++;
+	}
+
+	if (table[i].err_code == -EDEV_RECOVERED_ERROR)
+		rc = DEVICE_GOOD;
+	else if (table[i].sense == 0xFFFFFF && table[i].err_code == rc && msg)
+		*msg = (char*)(table[i].msg);
+
+	return rc;
+}
+
 struct supported_device {
 	char vendor_id[VENDOR_ID_LENGTH + 1];
 	char product_id[PRODUCT_ID_LENGTH + 1];
@@ -103,6 +146,61 @@ struct supported_device {
 };
 
 #define TAPEDRIVE(v, p, t, n) &(struct supported_device){ v, p, t, n }
+
+#define TAPE_FAMILY_MASK       (0xF000)
+#define TAPE_FAMILY_ENTERPRISE (0x1000)
+#define TAPE_FAMILY_LTO        (0x2000)
+#define TAPE_FAMILY_ARCHIVE    (0x4000)
+
+#define TAPE_FORMFACTOR_MASK   (0x0F00)
+#define TAPE_FORMFACTOR_FULL   (0x0100)
+#define TAPE_FORMFACTOR_HALF   (0x0200)
+
+#define TAPE_GEN_MASK          (0x00FF)
+
+#define IS_ENTERPRISE(type)    (type & TAPE_FAMILY_ENTERPRISE)
+#define IS_LTO(type)           (type & TAPE_FAMILY_LTO)
+#define IS_FULL_HEIGHT(type)   (type & TAPE_FORMFACTOR_FULL)
+#define IS_HALF_HEIGHT(type)   (type & TAPE_FORMFACTOR_HALF)
+#define DRIVE_FAMILY_GEN(type) (type & (TAPE_GEN_MASK | TAPE_FAMILY_MASK) )
+#define DRIVE_GEN(type)        (type & TAPE_GEN_MASK)
+
+enum {
+	VENDOR_UNKNOWN = 0,
+	VENDOR_IBM,
+	VENDOR_HP,
+};
+
+enum {
+	DRIVE_UNSUPPORTED = 0x0000, /* Unsupported drive */
+	DRIVE_LTO5        = 0x2105, /* Ultrium Gen 5 */
+	DRIVE_LTO5_HH     = 0x2205, /* Ultrium Gen 5 Half-High */
+	DRIVE_LTO6        = 0x2106, /* Ultrium Gen 6 */
+	DRIVE_LTO6_HH     = 0x2206, /* Ultrium Gen 6 Half-High */
+	DRIVE_LTO7        = 0x2107, /* Ultrium Gen 7 */
+	DRIVE_LTO7_HH     = 0x2207, /* Ultrium Gen 7 Half-High */
+	DRIVE_LTO8        = 0x2108, /* Ultrium Gen 8 */
+	DRIVE_LTO8_HH     = 0x2208, /* Ultrium Gen 8 Half-High */
+	DRIVE_LTO9        = 0x2109, /* Ultrium Gen 9 */
+	DRIVE_LTO9_HH     = 0x2209, /* Ultrium Gen 9 Half-High */
+	DRIVE_TS1140      = 0x1104, /* TS1140 */
+	DRIVE_TS1150      = 0x1105, /* TS1150 */
+	DRIVE_TS1155      = 0x5105, /* TS1155 */
+	DRIVE_TS1160      = 0x1106, /* TS1160 */
+};
+
+enum {
+	DRIVE_GEN_UNKNOWN = 0,
+	DRIVE_GEN_LTO5    = 0x2005,
+	DRIVE_GEN_LTO6    = 0x2006,
+	DRIVE_GEN_LTO7    = 0x2007,
+	DRIVE_GEN_LTO8    = 0x2008,
+	DRIVE_GEN_LTO9    = 0x2009,
+	DRIVE_GEN_JAG4    = 0x1004,
+	DRIVE_GEN_JAG5    = 0x1005,
+	DRIVE_GEN_JAG5A   = 0x5005,
+	DRIVE_GEN_JAG6    = 0x1006,
+};
 
 /* LTO cartridge type in mode page header */
 enum {
